Fixed bugs with mmap() code on the fault handler. It is working now but it's very inefficient and in some ways flawed. The following is needed:
1. Make sure that a given invocation of the fault handler
maps only the page that caused the fault (right now the entire
file is mapped). This should rid us of the need for most of the
local variables on the handler and also make it more efficient.

2.Make sure the case where the physical memory of an existing
mapping has been modified by the process holding the master mapping
(or any other process for that matter). The new process should get
a new physical page and the file data must be read again. This really
is part of #3.

3. Make sure that all mmap() flags and protection attributes are
handled correctly.