mmaped files are now loaded into physical memory on a page by page basis. Also the fault handler can now handle the case where an mmaped page is written to on it's first access in one shot. Previously it would take two faults, the first one would map the page and the 2nd one would make it writeable.
We now need to make sure that all the flags and protection settings
of mmap() are honored (right now mmaped files are made readable
and writeable by default).

We also need to make sure that the master mapping always contains
all the pages that have been loaded, otherwise we'll end up with
multiple copies in physical memory. Right now the only way for this
to happen would be if the master mapping is replaced (by a larger
mapping of the same file).

There are also several TODO anotations that need to be fixed.