Fixec ELF loader/mmap bugs.
Fixed bug in mmap where it wasn't properly handling the
situation when a mapping is not page aligned. Also fixed
fixed bug in elfld() where it incorrectly handling the case
where a section is not page aligned (it was aligning the address
and setting the mmaping incorrectly), now the mmaping is set on
the unaligned address and it's to mmap to handle the rest. What
this means is that it can now load ELF files made by any elf32
linker and not just mine.