Added script to create GRUB2 bootable image.
Added tools/scripts/make_image.sh. It creates a bootable
GRUB2 harddisk image. It then breaks into two images, one
for the partition table and one for the filesystem. This is
so that the build process can easily manipulate the filesystem
image with debugfs(3). Finally the filesystem image is compressed
so that it can be included in the repository.