Progress towards newlib port (working).
Newlib is now working. GCC still wont link to it
automatically unluess the library path is specified with
the -L option even to i386-hydrax-gcc -print-search-dirs
shows the right path on the list so I'm not sure why.

I'm also having problems linking pthread system call in
the same library as newlib so we'll have to go with a
separate pthreads library. For now the pthreads test
utility is not working.