More progress toward porting the C Library.
The parts of the C library that I have tested seem
to be working fine, however there seems to be problems
with some of the system calls that are not implemented
by the library. To complicate things I'm getting an error
in Kdevelop when I load the debugger and some breakpoints are
being hit but others are not even though I can tell that
the code is executed.