Fixed bugs with mmap and newlib port.
Fixed bug that caused new programs (after PID 1) to hang
because the process still had the mappings of the prior process.
Now the mappings are cleaned by kfree (which is called by execve
for each mmaping or allocation).

On newlib I fixed the errno.h header to play nice with it.

Note: There is a bug in newlib, if we exit a process by calling
exit() it calls the C library version which causes a segfault.
I believe this is because it tries to cleanup stuff that was
not initialized (since right now crt0 simply jumps to main()).