Working on Cygwin support.
I'm working on the 64 bit Cygwin distro on Windows XP x64
and everything works except building the mkinitrd utility.
The problem is that this utility needs to be built as 32-bit
but there appears to be a problem with the 32-bit compilers
on Cygwin64. I'm installing the 32-bit Cygwin right now,
hopefully everything will just work in it.

If it does I'll have to test on Windows 8.1 to make sure that
the problem is not related to Windows XP x64.